+++ src/ScottPlot4/ScottPlot/Plottable/AxisLine.cs
@@ -2,6 +2,7 @@
 using ScottPlot.Renderable;
 using System;
 using System.Drawing;
+using ScottPlot.SnapLogic;
 
 namespace ScottPlot.Plottable
 {
@@ -117,12 +118,7 @@ namespace ScottPlot.Plottable
         /// <summary>
         /// This function applies snapping logic while dragging
         /// </summary>
-        public Func<double, double> DragSnapX { get; set; } = (x) => x;
-
-        /// <summary>
-        /// This function applies snapping logic while dragging
-        /// </summary>
-        public Func<double, double> DragSnapY { get; set; } = (y) => y;
+        public ISnap2D DragSnap { get; set; } = new NoSnap2D();
 
         public AxisLine(bool isHorizontal)
         {
@@ -266,8 +262,10 @@ namespace ScottPlot.Plottable
             if (!DragEnabled)
                 return;
 
-            coordinateX = DragSnapX.Invoke(coordinateX);
-            coordinateY = DragSnapY.Invoke(coordinateY);
+            Coordinate original = new(coordinateX, coordinateY);
+            Coordinate snapped = DragSnap.Snap(original);
+            coordinateX = snapped.X;
+            coordinateY = snapped.Y;
 
             if (IsHorizontal)
             {