+++ src/ScottPlot4/ScottPlot/Plottable/DraggableMarkerPlot.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Linq;
-using System.Drawing;
-using System.Diagnostics;
-using ScottPlot.Drawing;
+using ScottPlot.SnapLogic;
 
 namespace ScottPlot.Plottable
 {
@@ -44,12 +41,7 @@ namespace ScottPlot.Plottable
         /// <summary>
         /// This function applies snapping logic while dragging
         /// </summary>
-        public Func<double, double> DragSnapX { get; set; } = (x) => x;
-
-        /// <summary>
-        /// This function applies snapping logic while dragging
-        /// </summary>
-        public Func<double, double> DragSnapY { get; set; } = (y) => y;
+        public ISnap2D DragSnap { get; set; } = new NoSnap2D();
 
         /// <summary>
         /// This event is invoked after the marker is dragged
@@ -67,8 +59,10 @@ namespace ScottPlot.Plottable
             if (!DragEnabled)
                 return;
 
-            coordinateX = DragSnapX(coordinateX);
-            coordinateY = DragSnapY(coordinateY);
+            Coordinate original = new(coordinateX, coordinateY);
+            Coordinate snapped = DragSnap.Snap(original);
+            coordinateX = snapped.X;
+            coordinateY = snapped.Y;
 
             if (coordinateX < DragXLimitMin) coordinateX = DragXLimitMin;
             if (coordinateX > DragXLimitMax) coordinateX = DragXLimitMax;