+++ src/ScottPlot4/ScottPlot/Plottable/IDraggable.cs
@@ -8,8 +8,7 @@ namespace ScottPlot.Plottable
         Cursor DragCursor { get; }
         bool IsUnderMouse(double coordinateX, double coordinateY, double snapX, double snapY);
         void DragTo(double coordinateX, double coordinateY, bool fixedSize);
-        public Func<double, double> DragSnapX { get; set; }
-        public Func<double, double> DragSnapY { get; set; }
+        public SnapLogic.ISnap2D DragSnap { get; set; }
         event EventHandler Dragged;
     }
 }