+++ src/ScottPlot4/ScottPlot/Snapping/1D/ISnap1D.cs
@@ -1,0 +1,6 @@
+﻿namespace ScottPlot.SnapLogic;
+
+public interface ISnap1D
+{
+    double Snap(double value);
+}