+++ src/ScottPlot4/ScottPlot/Snapping/1D/Integer1D.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace ScottPlot.SnapLogic;
+
+/// <summary>
+/// Snaps to the nearest integer position
+/// </summary>
+public class Integer1D : ISnap1D
+{
+    public double Snap(double value)
+    {
+        return Math.Round(value);
+    }
+}