+++ src/ScottPlot4/ScottPlot/Snapping/1D/NoSnap1D.cs
@@ -1,0 +1,12 @@
+﻿namespace ScottPlot.SnapLogic;
+
+/// <summary>
+/// Always returns the original value (snapping disabled)
+/// </summary>
+public class NoSnap1D : ISnap1D
+{
+    public double Snap(double value)
+    {
+        return value;
+    }
+}