+++ src/ScottPlot4/ScottPlot/Snapping/2D/ISnap2D.cs
@@ -1,0 +1,6 @@
+﻿namespace ScottPlot.SnapLogic;
+
+public interface ISnap2D
+{
+    Coordinate Snap(Coordinate value);
+}