+++ src/ScottPlot4/changelog.md
@@ -9,10 +9,10 @@ _not yet published on NuGet..._
 * Draggable Marker Plots: Improved drag behavior when drag limits are in use (#1970) _Thanks @xmln17_
 * Signal Plot: Added support for plotting `byte` arrays (#1945)
 * Axis Line: Added properties to customize alignment of position labels (#1972) _Thanks @hamhub7_
-* Draggables: Added `DragSnapX` and `DragSnapY` functions to let users customize snapping (#2006, #2007) _Thanks @Agorath_
 * Plot: MatchAxis no longer modifies limits of unintended axes (#1980) _Thanks @PlayCreatively_
 * Plot: Improved error reporting for invalid axis limits (#1994) _Thanks @Xerxes004_
 * Signal Plot: Improved `GetPointNearestX()` accuracy for plots with high zoom (#1987, #2019, #2020) _Thanks @dhgigisoave_
+* Draggable: Added custom functions to permit 1D and 2D snapping (#2006, #2007, #2022) _Thanks @Agorath_
 
 ## ScottPlot 4.1.52
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2022-07-09_